docs: remove compatibility table
This information is either out-of-date or confusing, since it mentions Ember and Ember Data versions but not which versions of this package work with those combinations. Reading the `README`, one might think that this package actively is tested against and works with each of those Ember/Ember Data pairings, which is not really the case.